Update: DynamicSidebar 분리
기존에 하나로 합쳐져있던 사이드바 코드를 모드별로 분리해서 파일을 만들었습니다
